Fix a copy/paste error that truncated proper values for audio deviation.

+++ commands.h
@@ -107,7 +107,7 @@ public:
 	bool saveParametersToVast();
 
 private:
-	uint8_t mAudioDeviation;// = 0;
+	uint32_t mAudioDeviation;// = 0;
 };
 
 class VastPilotLmrRds : public VastCommand
